feat: allow editing robots
This is important now that we allow adding robots without a unique
serial number.
For such cases we only store the usb port that was selected when adding
the robot. But since this port can change when reconnecting the robot or
when rebooting a system the user could end up with their robot living on
a different port, and thus showing up as disconnected.
