+++ application/ui/src/features/robots/robot-form/form.tsx
@@ -1,6 +1,4 @@
-import { useRef, type FormEvent } from 'react';
-
-import { Button, Divider, Flex, Form, Heading, Icon, Item, Picker, TextField } from '@geti-ui/ui';
+import { Button, Flex, Heading, Icon, Item, Picker, TextField } from '@geti-ui/ui';
 import { ChevronLeft } from '@geti-ui/ui/icons';
 
 import { useProjectId } from '../../../features/projects/use-project';
@@ -10,9 +8,8 @@ import { SO101FormFields } from './catalog/so101';
 import { WidowxAIFormFields } from './catalog/widowxai';
 import { BiManualWidowxAIFormFields } from './catalog/widowxai-bimanual';
 import { useRobotForm, useRobotFormFields, useSetRobotForm } from './provider';
-import { SubmitNewRobotButton } from './submit-new-robot-button';
 
-const RobotType = () => {
+export const RobotType = () => {
     const { activeType } = useRobotForm();
     const { setActiveType } = useSetRobotForm();
 
@@ -38,11 +35,11 @@ const RobotType = () => {
     );
 };
 
-const FormFields = ({ robotType }: { robotType: SchemaRobotType }) => {
-    const { formData, updateField } = useRobotFormFields();
+export const FormFields = () => {
+    const { formData, updateField, activeType } = useRobotFormFields();
 
     let formFields = null;
-    switch (robotType) {
+    switch (activeType) {
         case 'SO101_Follower':
         case 'SO101_Leader':
             formFields = <SO101FormFields />;
@@ -73,49 +70,22 @@ const FormFields = ({ robotType }: { robotType: SchemaRobotType }) => {
     );
 };
 
-export const RobotForm = ({ heading = 'Add new robot', submitButton = <SubmitNewRobotButton /> }) => {
+export const RobotFormHeading = ({ heading }: { heading: string }) => {
     const { project_id } = useProjectId();
 
-    const { activeType } = useRobotForm();
-
-    const submitContainerRef = useRef<HTMLDivElement>(null);
-
-    // Make Enter behave like clicking the submit button. A disabled button ignores
-    // the click, so validation still applies.
-    const handleSubmit = (event: FormEvent<HTMLFormElement>) => {
-        event.preventDefault();
-        submitContainerRef.current?.querySelector('button')?.click();
-    };
-
     return (
-        <Flex direction='column' gap='size-200'>
-            <Flex alignItems={'center'} gap='size-200'>
-                <Button
-                    href={paths.project.robots.index({ project_id })}
-                    variant='secondary'
-                    UNSAFE_style={{ border: 'none' }}
-                >
-                    <Icon>
-                        <ChevronLeft color='white' fill='white' />
-                    </Icon>
-                </Button>
+        <Flex alignItems={'center'} gap='size-200'>
+            <Button
+                href={paths.project.robots.index({ project_id })}
+                variant='secondary'
+                UNSAFE_style={{ border: 'none' }}
+            >
+                <Icon>
+                    <ChevronLeft color='white' fill='white' />
+                </Icon>
+            </Button>
 
-                <Heading>{heading}</Heading>
-            </Flex>
-            <Divider orientation='horizontal' size='S' />
-            {/* Prevent native form submission, which reloads the page and discards form state.
-                Advancing to the next step is handled by the submit button's onPress. */}
-            <Form onSubmit={handleSubmit}>
-                <Flex direction='column' gap='size-200'>
-                    <Flex direction='column' gap='size-200' width='100%'>
-                        <RobotType />
-
-                        <FormFields robotType={activeType} />
-                    </Flex>
-                    <Divider orientation='horizontal' size='S' />
-                    <div ref={submitContainerRef}>{submitButton}</div>
-                </Flex>
-            </Form>
+            <Heading>{heading}</Heading>
         </Flex>
     );
 };
