+++ application/ui/src/routes/robots/new.tsx
@@ -2,9 +2,8 @@ import { Suspense } from 'react';
 
 import { Flex, Grid, Loading, minmax, View } from '@geti-ui/ui';
 
-import { RobotForm } from '../../features/robots/robot-form/form';
+import { CreateRobotForm } from '../../features/robots/robot-form/create-form';
 import { Preview } from '../../features/robots/robot-form/preview';
-import { SubmitNewRobotButton } from '../../features/robots/robot-form/submit-new-robot-button';
 
 const CenteredLoading = () => {
     return (
@@ -19,7 +18,7 @@ export const New = () => {
         <Grid areas={['robot controls']} columns={[minmax('size-6000', 'auto'), '1fr']} height={'100%'}>
             <View gridArea='robot' backgroundColor={'gray-100'} padding='size-400'>
                 <Suspense fallback={<CenteredLoading />}>
-                    <RobotForm submitButton={<SubmitNewRobotButton />} />
+                    <CreateRobotForm />
                 </Suspense>
             </View>
             <View gridArea='controls'>
